+++ src/mongocxx/private/libmongoc.hh
@@ -25,7 +25,8 @@ MONGOCXX_INLINE_NAMESPACE_BEGIN
 namespace libmongoc {
 
 #ifdef MONGOCXX_TESTING
-#define MONGOCXX_LIBMONGOC_SYMBOL(name) extern mongocxx::mock::mock<decltype(&mongoc_##name)>& name;
+#define MONGOCXX_LIBMONGOC_SYMBOL(name) \
+    extern mongocxx::test_util::mock<decltype(&mongoc_##name)>& name;
 #include "libmongoc_symbols.hh"
 #undef MONGOCXX_LIBMONGOC_SYMBOL
 #else
